Remove package-lock.json file from repo.
Since this is a library, the lock file won't take effect on installation
in other projects.

This means that depfu will only send PRs when there is new versions of a
depedency that is not satisfied by our semver-range in package.json, but
it will still run CI jobs for `dependencies` covered by our semver
range, but will only create PRs when CI breaks.

An alternate strategy would have been to use a `.npmrc` with a line that
disables the lock file (`package-lock=false`).